app/src/main/java/com/example/soccerapps/presenter/NextEventTeamPresenter.kt
package com.example.soccerapps.presenter

import com.example.soccerapps.api.ApiRepository
import com.example.soccerapps.api.TheSportDBApi
import com.example.soccerapps.model.EventResponse
import com.example.soccerapps.view.interfaces.NextTeamView
import com.example.soccerapps.view.util.CoroutineContextProvider
import com.google.gson.Gson
import kotlinx.coroutines.GlobalScope
import kotlinx.coroutines.launch

class NextEventTeamPresenter(
    private val nextTeamView: NextTeamView,
    private val apiRepository: ApiRepository,
    private val gson: Gson,
    private val context: CoroutineContextProvider = CoroutineContextProvider()
) {
    fun getEvenTeamList(teamId: String?) {
        nextTeamView.showLoading()
        GlobalScope.launch(context.main) {
            val data =
                gson.fromJson(
                    apiRepository.doRequestAsync(
                        TheSportDBApi.getNextEventTeam(teamId)
                    ).await(), EventResponse::class.java
                )
            nextTeamView.hideLoading()
            nextTeamView.showEventTeamList(data.events)
        }
    }
}